+++ example/lib/main.dart
@@ -70,6 +70,7 @@ class _MyAppState extends State<MyApp> {
   String? selectedValueUpdateFromOutsideThePlugin;
   dynamic selectedValueSingleDialogPaged;
   dynamic selectedValueSingleDialogPagedFuture;
+  dynamic selectedValueSingleDialogFuture;
   ExampleNumber? selectedNumber;
   List<int> selectedItemsMultiDialog = [];
   List<int> selectedItemsMultiCustomDisplayDialog = [];
@@ -100,6 +101,8 @@ class _MyAppState extends State<MyApp> {
   Function? openDialog;
 
   PointerThisPlease<int> currentPage = PointerThisPlease<int>(1);
+
+  bool noResult = false;
 
   @override
   void initState() {
@@ -1482,6 +1485,119 @@ class _MyAppState extends State<MyApp> {
         },
         menuConstraints: BoxConstraints.tight(Size.fromHeight(350)),
       ),
+      "Single dialog custom empty list": SearchChoices.single(
+        items: items,
+        value: selectedValueSingleDialog,
+        hint: "Select one",
+        searchHint: "Select one",
+        onChanged: (value) {
+          setState(() {
+            selectedValueSingleDialog = value;
+          });
+        },
+        isExpanded: true,
+        emptyListWidget: (String keyword) =>
+            "No result with the \"$keyword\" keyword",
+      ),
+      "Single dialog future custom empty list": SearchChoices.single(
+        value: selectedValueSingleDialogFuture,
+        hint: kIsWeb ? "Example not for web" : "Select one capital",
+        searchHint: "Search capitals",
+        onChanged: kIsWeb
+            ? null
+            : (value) {
+                setState(() {
+                  selectedValueSingleDialogFuture = value;
+                });
+              },
+        isExpanded: true,
+        selectedValueWidgetFn: (item) {
+          return (Center(
+              child: Card(
+                  shape: RoundedRectangleBorder(
+                    borderRadius: BorderRadius.circular(4),
+                    side: BorderSide(
+                      color: Colors.grey,
+                      width: 1,
+                    ),
+                  ),
+                  margin: EdgeInsets.all(1),
+                  child: Padding(
+                    padding: const EdgeInsets.all(6),
+                    child: Text(item["capital"]),
+                  ))));
+        },
+        futureSearchFn: (String? keyword, String? orderBy, bool? orderAsc,
+            List<Tuple2<String, String>>? filters, int? pageNb) async {
+          String filtersString = "";
+          int i = 1;
+          filters?.forEach((element) {
+            // This example doesn't have any futureSearchFilterOptions parameter, thus, this loop will never run anything.
+            filtersString += "&filter" +
+                i.toString() +
+                "=" +
+                element.item1 +
+                "," +
+                element.item2;
+            i++;
+          });
+          Response response = await get(Uri.parse(
+                  "https://searchchoices.jod.li/exampleList.php?page=${pageNb ?? 1},10${orderBy == null ? "" : "&order=" + orderBy + "," + (orderAsc ?? true ? "asc" : "desc")}${(keyword == null || keyword.isEmpty) ? "" : "&filter=capital,cs," + keyword}$filtersString"))
+              .timeout(Duration(
+            seconds: 10,
+          ));
+          if (response.statusCode != 200) {
+            throw Exception("failed to get data from internet");
+          }
+          dynamic data = jsonDecode(response.body);
+          int nbResults = data["results"];
+          List<DropdownMenuItem> results = (data["records"] as List<dynamic>)
+              .map<DropdownMenuItem>((item) => DropdownMenuItem(
+                    value: item,
+                    child: Card(
+                      shape: RoundedRectangleBorder(
+                        borderRadius: BorderRadius.circular(4),
+                        side: BorderSide(
+                          color: Colors.blue,
+                          width: 1,
+                        ),
+                      ),
+                      margin: EdgeInsets.all(1),
+                      child: Padding(
+                        padding: const EdgeInsets.all(6),
+                        child: Text(
+                            "${item["capital"]} - ${item["country"]} - ${item["continent"]} - pop.: ${item["population"]}"),
+                      ),
+                    ),
+                  ))
+              .toList();
+          return (Tuple2<List<DropdownMenuItem>, int>(results, nbResults));
+        },
+        emptyListWidget: () => Text(
+          "No result",
+          style: TextStyle(
+            fontStyle: FontStyle.italic,
+            color: Colors.grey,
+          ),
+        ),
+      ),
+      "Single dialog onTap": SearchChoices.single(
+        items: items,
+        value: selectedValueSingleDialog,
+        hint: "Select one",
+        searchHint: "Select one",
+        onChanged: (value) {
+          setState(() {
+            selectedValueSingleDialog = value;
+          });
+        },
+        isExpanded: true,
+        onTap: () {
+          setState(() {
+            selectedValueSingleDialog = null;
+          });
+        },
+      ),
     };
 
     return MaterialApp(
@@ -1561,7 +1677,14 @@ class _MyAppState extends State<MyApp> {
                                     )))));
                       })
                       .values
-                      .toList(),
+                      .toList()
+                        ..add(
+                          Center(
+                            child: SizedBox(
+                              height: 500,
+                            ),
+                          ),
+                        ), //prevents scrolling issues at the end of the list of Widgets
                 ),
               ),
             ),